+++ app/build.gradle
@@ -27,7 +27,7 @@ android {
     }
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.example.helloworldapp"
+        applicationId "com.example.natalevichmvapp"
         minSdkVersion 22
         targetSdkVersion 28
         versionCode 1
